Add samples to a root every 50 millis
Rather than dumping all the samples after the end sample key press, this
adds a timer to collect the samples into a root every 20th of a second.
When the key is pressed, the latest samples are collected and the root is
dumped. If something big is happening, the sample buffer is overwhelmed before
the getSamples call can go through.

There's a flash.sampler.setSamplerCallback function that's supposed to
call through when the buffer is nearly full, but it doesn't actually
compile in flex 4.1 even though it's documented for it. Zell found that
it's been replaced with

void Sampler::setCallback(ScriptObject* callback)
{
        (void)callback;
        // FIXME: bug 614529
        //         //this->callback = callback;
}

Smooth, Adobe.